This is messy but to summarize ive moved all the code from the previous session to doc folder and created components for the game. In the following commits the signature of the methods will be finished'
