+++ Pipfile
@@ -7,6 +7,7 @@ name = "pypi"
 pygame = "*"
 click = "*"
 jam-session-pygame = {editable = true, path = "."}
+pygame-gui = "*"
 
 [dev-packages]
