+++ docs/lib_old/entities/asset.py
