+++ docs/lib_old/entities/game_state.py
