+++ docs/lib_old/entities/player.py
