+++ docs/lib_old/entities/stage.py
