+++ docs/lib_old/to_trash.py
