+++ jam_session/lib/components/event_handler.py
@@ -1,0 +1,2 @@
+class DefaultEventHandler:
+    pass
