--- jam_session/lib/components/game_interface.py
@@ -1,4 +1,0 @@
-class GameInterface:
-    """
-    """
-    pass
